Rename package runner to execution
