+++ src/main/kotlin/com/github/jozott00/wokwiintellij/execution/WokwiProcessHandler.kt
@@ -1,4 +1,4 @@
-package com.github.jozott00.wokwiintellij.runner
+package com.github.jozott00.wokwiintellij.execution
 
 import com.github.jozott00.wokwiintellij.simulator.WokwiSimulatorListener
 import com.intellij.execution.process.ProcessHandler
