+++ src/main/kotlin/com/github/jozott00/wokwiintellij/execution/macros/GdbServerMacro.kt
@@ -1,4 +1,4 @@
-package com.github.jozott00.wokwiintellij.runner.macros
+package com.github.jozott00.wokwiintellij.execution.macros
 
 import com.github.jozott00.wokwiintellij.toml.WokwiConfigProcessor
 import com.intellij.ide.macro.Macro
